central_server/requirements.txt
# Central Server - Instagram Tools Multi-Tenant System

fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0  # For async SQLite
aiomysql>=0.2.0  # For async MySQL
PyMySQL>=1.1.0  # Dependency for aiomysql
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
tenacity>=8.2.0
